requirements.txt
affine==2.4.0
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
fiona==1.9.5
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.8.0
multidict==6.0.4
numpy==1.24.4
Pillow==10.1.0
pyparsing==3.1.1
rasterio==1.3.9
six==1.16.0
snuggs==1.4.7
yarl==1.9.2
zipp==3.17.0
